joyle/requirements.txt
backports-abc==0.5
certifi==2018.1.18
dj-database-url==0.5.0
Django==1.11.10
django-redis==4.9.0
django-shortuuidfield==0.1.3
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
PyJWT==1.5.3
pytz==2017.3
redis==2.10.6
shortuuid==0.5.0
singledispatch==3.4.0.3
six==1.11.0
tornado==4.5.3
whitenoise==3.3.1
psycopg2==2.7.4
